[WIP] Attempt to fix arc layer props
The mockCredits object have the correct data to draw arcs in arc layer. It's being passed as a promise hence it's throwing errors. Will work on that next.
The implementation of BrushingExtension's example is understandable but probably needs modification of the object. That's why I decided to try make the arc layer work first (the brushing layer also uses arc layer).