Update mock data object structure and apply brushing
dataLoading needs to be improved because it takes a while now.
To do this, need to change the hexagon data creation. See the data loaders functions and merge into a single data loading.
The mock transaction data can be created from the csv of Singapore locations directly and hexagonData can be deprecated after that.
This should improve the load time by a lot.